@@ -1,11 +1,11 @@
 name: interactiveviewer_gallery
 description: Image and video preview component that supports zoom and drag
-version: 0.3.0
+version: 0.4.0
 author: qq326646683<326646683@qq.com>
 homepage: https://github.com/qq326646683/interactiveviewer_gallery
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
